refactor: move config structs to config.rs
