+++ src/lsm.rs
@@ -7,6 +7,7 @@ use std::sync::atomic::AtomicU64;
 use bytes::Bytes;
 
 use crate::{
+    config::{MemtableConfig, WalConfig},
     memtable::Memtable,
     wal::{Wal, WalEntry},
 };
@@ -31,35 +32,6 @@ pub struct Lsm {
     l2_files: Vec<u64>,
 
     working_directory: PathBuf,
-}
-
-#[derive(Debug, Clone)]
-pub struct WalConfig {
-    id: u64,
-    max_size: u64,
-    log_directory: PathBuf,
-}
-
-impl WalConfig {
-    pub fn new(id: u64, max_size: u64, log_directory: PathBuf) -> Self {
-        Self {
-            id,
-            max_size,
-            log_directory,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MemtableConfig {
-    id: u64,
-    max_size: u64,
-}
-
-impl MemtableConfig {
-    pub fn new(id: u64, max_size: u64) -> Self {
-        Self { id, max_size }
-    }
 }
 
 impl Lsm {
